+++ server.js
@@ -4,12 +4,14 @@ const fs = require('fs');
 const isBot = require('isbot');
 const puppeteer = require('puppeteer');
 const express = require('express');
-const app = express();
 const path = require('path');
+const btoa = require('btoa');
+const fetch = require('node-fetch');
 
 const port = process.env.port || 8000;
 const host = process.env.port || "localhost";
 
+const app = express();
 let browser = null;
 
 const replaceLinksIfBot = (html) => html.replace(/href="\/#/g, 'href="/');
@@ -92,6 +94,120 @@ router.get('/address/:item', async (req, res) =>
 
 router.get('/block/:item', async (req, res) => 
   loadPage(res, req, `http://127.0.0.1:${port}/?isbot=${isBot(req.header['user-agent'])}#block/${req.params.item}`));
+
+router.get('/sitemap.xml', async (req, res) => {
+  const tokens_query = {
+    "v": 3,
+    "q": {
+      "db": ["t"],
+      "find": {
+        "tokenStats.approx_txns_since_genesis": {
+          "$gte": 100
+        }
+      },
+      "sort": {
+        "tokenStats.approx_txns_since_genesis": -1
+      },
+      "project": {
+        "_id": 0,
+        "tokenDetails.tokenIdHex": 1,
+        "tokenStats": 1
+      },
+      "limit": 10000
+    }
+  };
+  
+  const addresses_query = {
+    "v": 3,
+    "db": ["g"],
+    "q": {
+      "aggregate": [
+        {
+          "$match": {}
+        },
+        {
+          "$limit": 1000000
+        },
+        {
+          "$unwind": "$graphTxn.outputs"
+        },
+        {
+          "$group": {
+            "_id": "$graphTxn.outputs.address",
+            "cnt": {
+              "$sum": 1
+            }
+          }
+        },
+        {
+          "$sort": {
+            "cnt": -1
+          }
+        }
+      ],
+      "sort": {
+        "cnt": -1
+      },
+      "limit": 10000
+    }
+  };
+
+  const transactions_query = {
+    "v": 3,
+    "q": {
+      "db": ["c"],
+      "find": {},
+      "sort": {
+        "blk.i": -1
+      },
+      "project": {
+        "_id": -1,
+        "tx.h": 1
+      },
+      "limit": 10000
+    }
+  };
+
+
+  let response = "";
+  
+  const slpdbQuery = (query) => fetch(`https://slpdb.fountainhead.cash/q/${btoa(JSON.stringify(query))}`)
+  .then((data) => data.json());
+  
+  response += `<?xml version="1.0" encoding="UTF-8"?>
+  <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+  <url><loc>https://simpleledger.info/</loc></url>
+  <url><loc>https://simpleledger.info/alltokens</loc></url>
+  <url><loc>https://simpleledger.info/dividend</loc></url>
+  <url><loc>https://simpleledger.info/block/mempool</loc></url>`;
+  
+  
+  return slpdbQuery(tokens_query)
+  .then((data) => {
+    for (let m of data.t) {
+      response += `<url><loc>https://simpleledger.info/token/${m.tokenDetails.tokenIdHex}</loc></url>`;
+    }
+  
+    return slpdbQuery(addresses_query)
+    .then((data) => {
+      for (let m of data.g) {
+        response += `<url><loc>https://simpleledger.info/address/${m._id}</loc></url>`;
+      }
+
+      return slpdbQuery(transactions_query)
+      .then((data) => {
+        for (let m of data.c) {
+          response += `<url><loc>https://simpleledger.info/tx/${m.tx.h}</loc></url>`;
+        }
+  
+        response += `</urlset>`;
+
+        res.set('Content-Type', 'application/xml');
+        return res.send(response);
+      });
+    });
+  });
+});
 
 app.use('/', router);
 app.use('/', express.static('public'));
